% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rburrxii.R
\name{rburrxii}
\alias{rburrxii}
\title{Generates a random sample from a Burr XII distribution}
\usage{
rburrxii(n, shape1, shape2)
}
\arguments{
\item{n}{number of observations.}

\item{shape1, shape2}{shape parameters. Must be strictly positive.}
}
\value{
A vector containing data distributed according to the Burr XII
distribution with a sample size specified by \code{n}, and with shape
parameters specified by \code{shape1} and \code{shape2}.
}
\description{
This function is designed to generate random data following a Burr XII
distribution with parameters \code{shape1} and \code{shape2}.
}
\details{
The Burr XII distribution with parameters \code{shape1} = \eqn{\gamma}
and \code{shape2} = \eqn{\alpha} has the probability density function:
\deqn{f_{X}(x) = \alpha\gamma\dfrac{x^{\alpha-1}}{(1+x^\alpha)^{\gamma+1}}}
for \eqn{x>0}, \eqn{\gamma>0} and \eqn{\alpha>0}.
}
\examples{
rburrxii(10, 2, 3)
rburrxii(100, 4, 16)

}
