% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rloglogis.R
\name{rloglogis}
\alias{rloglogis}
\title{Generates a random sample from a Log-logistic distribution}
\usage{
rloglogis(n, scale, shape)
}
\arguments{
\item{n}{number of observations.}

\item{scale}{scale parameter. Must be strictly positive.}

\item{shape}{shape parameter. Must be strictly positive.}
}
\value{
A vector containing data distributed according to
the Log-logistic distribution with a sample size specified by \code{n}, and with scale
and shape parameters specified by \code{scale} and \code{shape}, respectively.
}
\description{
This function is designed to generate random data following
a Log-logistic distribution with parameters \code{scale} and \code{shape}.
}
\details{
The Log-logistic distribution with parameters \code{scale} = \eqn{\alpha}
and \code{shape} = \eqn{\beta} has the probability density function:
\deqn{f_{X}(x) = \dfrac{(\beta/\alpha)(x/\alpha)^{\beta-1}}{(1 + (x/\alpha)^{\beta})^{2}}}
for \eqn{x\geq 0}, \eqn{\theta>0} and \eqn{\alpha>0}.
}
\examples{
rloglogis(10, 5, 2)
rloglogis(100, 10, 6)

}
