% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rparetoi.R
\name{rparetoi}
\alias{rparetoi}
\title{Generates a random sample from a Pareto I distribution}
\usage{
rparetoi(n, scale, shape)
}
\arguments{
\item{n}{number of observations.}

\item{scale}{scale parameter. Must be strictly positive.}

\item{shape}{shape parameter. Must be strictly positive.}
}
\value{
A vector containing data distributed according to the Pareto I
distribution with a sample size specified by \code{n}, and with scale
and shape parameters specified by \code{scale} and \code{shape}, respectively.
}
\description{
This function is designed to generate random data following a Pareto I
distribution with parameters \code{scale} and \code{shape}.
}
\details{
The Pareto I distribution with parameters \code{scale} = \eqn{\theta}
and \code{shape} = \eqn{\alpha} has the probability density function:
\deqn{f_{X}(x) = \dfrac{\alpha\theta^{\alpha}}{x^{\alpha+1}}}
for \eqn{x\geq \theta}, \eqn{\theta>0} and \eqn{\alpha>0}.
}
\examples{
rparetoi(10, 5, 2)
rparetoi(100, 10, 6)

}
